+++ vizro-core/examples/scratch_dev/app.py
@@ -1,80 +1,46 @@
-# Vizro is an open-source toolkit for creating modular data visualization applications.
-# check out https://github.com/mckinsey/vizro for more info about Vizro
-# and checkout https://vizro.readthedocs.io/en/stable/ for documentation.
+# # Vizro is an open-source toolkit for creating modular data visualization applications.
+# # check out https://github.com/mckinsey/vizro for more info about Vizro
+# # and checkout https://vizro.readthedocs.io/en/stable/ for documentation.
 
+import vizro.models as vm
 import vizro.plotly.express as px
 from vizro import Vizro
-import vizro.models as vm
-from vizro.models.types import capture
+from vizro.tables import dash_ag_grid, dash_data_table
 
+tips = px.data.tips()
 
-@capture("action")
-def action_function_output_list(button_number_of_clicks, _controls):
-    return f"Button clicked {button_number_of_clicks} times.", f"Controls: {str(_controls)}"
-
-
-@capture("action")
-def action_function_output_dict(button_number_of_clicks, _controls):
-    return {
-        "card-id-2": f"Controls: {str(_controls)}.",
-        "card-id-1": f"Button clicked {button_number_of_clicks} times.",
-    }
-
-
-@capture("action")
-def legacy_action_function_output_dict(button_number_of_clicks):
-    return {
-        "card-id-2": f"Button clicked {button_number_of_clicks} times.",
-        "card-id-1": f"Button clicked {button_number_of_clicks} times.",
-    }
-
-
-# TODO: Enable all three cases
-action_output_list = vm.Action(
-    function=action_function_output_list("button-id.n_clicks"),
-    outputs=["card-id-1.children", "card-id-2.children"],
-)
-action_output_dict = vm.Action(
-    function=action_function_output_dict("button-id.n_clicks"),
-    outputs={
-        "card-id-1": "card-id-1.children",
-        "card-id-2": "card-id-2.children",
-    },
-)
-action_output_dict_legacy = vm.Action(
-    function=legacy_action_function_output_dict(),
-    inputs=["button-id.n_clicks"],
-    outputs={
-        "card-id-1": "card-id-1.children",
-        "card-id-2": "card-id-2.children",
-    },
-)
-
-
-page = vm.Page(
-    title="Title",
-    layout=vm.Flex(),
+page1 = vm.Page(
+    title="LAYOUT_FLEX_WRAP_AND_AG_GRID",
+    layout=vm.Flex(wrap=True),
     components=[
-        vm.Graph(figure=px.scatter(px.data.iris(), x="sepal_length", y="petal_width", color="species")),
-        vm.Button(
-            id="button-id",
-            text="Click me",
-            actions=[
-                # action_output_list,
-                # action_output_dict,
-                action_output_dict_legacy,
-            ],
-        ),
-        vm.Card(
-            id="card-id-1",
-            text="Click button to update me",
-        ),
-        vm.Card(id="card-id-2", text="Click button to update me"),
+        vm.AgGrid(id=f"outer_id_{i}", figure=dash_ag_grid(tips, id=f"inner_id_{i}", style={"width": 1000}))
+        # vm.AgGrid(id=f"outer_id_{i}", figure=dash_ag_grid(tips, id=f"qwer", style={"width": 1000}))
+        for i in range(3)
     ],
-    controls=[vm.Filter(column="species", selector=vm.RadioItems())],
 )
 
-dashboard = vm.Dashboard(pages=[page])
+page2 = vm.Page(
+    title="LAYOUT_FLEX_GAP_AND_TABLE",
+    layout=vm.Flex(gap="40px"),
+    # components=[vm.Table(figure=dash_data_table(tips, style_table={"width": "1000px"})) for i in range(3)],
+    components=[
+        vm.Table(figure=dash_data_table(tips, id=f"qwert_{i}", style_table={"width": "1000px"})) for i in range(3)
+    ],
+)
+
+page3 = vm.Page(
+    title="cross_id",
+    layout=vm.Flex(gap="40px"),
+    components=[
+        vm.Table(figure=dash_data_table(tips, id="qwert", style_table={"width": "1000px"})),
+        vm.AgGrid(figure=dash_ag_grid(tips, id="qwert", style={"width": 1000})),
+    ],
+)
+
+dashboard = vm.Dashboard(pages=[page1, page2, page3])
+
 
 if __name__ == "__main__":
     Vizro().build(dashboard).run()
+
+## TO CHECK: also for figure and/or Graph?