+++ vizro-core/tests/unit/vizro/models/_action/test_action.py
@@ -297,6 +297,26 @@ class TestActionInputs:
         action = Action(function=action_function(**inputs))
         assert action._transformed_inputs == expected_transformed_inputs
 
+    @pytest.mark.parametrize(
+        "input",
+        [
+            ["component.property"],
+            1,
+            None,
+            "",
+            "component",
+            "component.",
+            ".property",
+            "component..property",
+            "component_property",
+            "component.property.property",
+        ],
+    )
+    @pytest.mark.xfail(reason="Validation will only be performed once legacy actions are removed")
+    def test_runtime_inputs_invalid(self, input):
+        with pytest.raises(ValidationError):
+            Action(function=action_with_one_arg(input))._transformed_inputs
+
 
 class TestBuiltinRuntimeArgs:
     """Test the actual values of the runtime args are correct in a real scenario."""
@@ -336,6 +356,9 @@ class TestActionOutputs:
     @pytest.mark.parametrize(
         "outputs",
         [
+            "component.property",
+            1,
+            None,
             [""],
             ["component"],
             ["component."],
@@ -351,10 +374,11 @@ class TestActionOutputs:
             {"output_1": "component_property"},
             {"output_1": "component.property.property"},
             {"output_1": "component.property", "output_2": ""},
+            {1: "component.property"},
         ],
     )
     def test_outputs_invalid(self, outputs):
-        with pytest.raises(ValidationError, match="String should match pattern"):
+        with pytest.raises(ValidationError):
             Action(function=action_with_no_args(), outputs=outputs)