+++ vizro-core/tests/unit/vizro/models/_components/form/test_dropdown.py
@@ -274,7 +274,7 @@ class TestDropdownBuild:
             description=Tooltip(text="Test description", icon="info", id="info"),
         ).build()
 
-        description = [
+        expected_description = [
             html.Span("info", id="info-icon", className="material-symbols-outlined tooltip-icon"),
             dbc.Tooltip(
                 children=dcc.Markdown("Test description", className="card-text"),
@@ -286,7 +286,7 @@ class TestDropdownBuild:
 
         expected_dropdown = html.Div(
             [
-                dbc.Label(["Title", *description], html_for="dropdown_id"),
+                dbc.Label(["Title", *expected_description], html_for="dropdown_id"),
                 dcc.Dropdown(
                     id="dropdown_id",
                     options=[